+++ src/onegov.core/onegov/core/orm/mixins/__init__.py
@@ -1,3 +1,4 @@
+from onegov.core.orm.mixins.content import ContentMixin
 from onegov.core.orm.mixins.timestamp import TimestampMixin
 
-__all__ = ['TimestampMixin']
+__all__ = ['ContentMixin', 'TimestampMixin']